setup autodocs with custom template
